opnbadge_analysis folder to correct place
Rechecked the openbadge_analysis directory that I extracted from the side-packages. So now the openbadge_analysis directory should be going to the right place (under openbadge-analysis/).
