--- openbadge_analysis_2/__init__.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-
-__author__ = 'Oren Lederman'
-__email__ = 'orenled@mit.edu'
-__version__ = '0.4.4'
-
-__all__ = ["core", "data", "preprocessing", "visualization", "analysis"]
-
