--- openbadge_analysis_2/analysis/__init__.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from .gatherings_cores import extract_groups, gather_groups, extract_cores
-
