--- openbadge_analysis_2/visualization/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-
-
-from .embedding import rssi_to_distance, members_2d_embedding, members_2d_embedding_with_beacons
-
-__all__ = ["audio_visualization" ]
